LS1043 : Fix Integer Overflow compilation error introduced inadvertently in commit 17d39cabee3eab5935c60d7e96c272c4e1e8c4e4 The addition of CONFIG_SYS_LS_PCIE_MEM_PHYS_OFF and CONFIG_SYS_LS_PCIE_MEM_PHYS_OFF causing the signed bit of 32 bit integer to set, which causes integer overflow error. Therfore typecasting CONFIG_SYS_LS_PCIE_MEM_PHYS_OFF to unsigned 32 bit integer before adding
Contributed-under: TianoCore Contribution Agreement 1.0

+++ LS1043aRdbPkg/Library/PciHostBridgeLib/PciRbLib.c
@@ -115,7 +115,8 @@ RootBridgeIoMemRW (
   }
 
   if ((Address < CONFIG_SYS_LS_PCIE_MEM_PHYS_OFF) ||
-      (Address >= (CONFIG_SYS_LS_PCIE_MEM_PHYS_OFF + CONFIG_SYS_LS_PCIE_MEM_SIZE)) {
+      (Address >= ((UINT32)CONFIG_SYS_LS_PCIE_MEM_PHYS_OFF +
+                   CONFIG_SYS_LS_PCIE_MEM_SIZE))) {
     return EFI_INVALID_PARAMETER;
   }
 
@@ -254,7 +255,8 @@ RootBridgeIoPciRW (
   }
 
   if ((Address < CONFIG_SYS_LS_PCIE_IO_PHYS_OFF) ||
-      (Address >= (CONFIG_SYS_LS_PCIE_IO_PHYS_OFF + CONFIG_SYS_LS_PCIE_IO_SIZE)) {
+      (Address >= ((UINT32)CONFIG_SYS_LS_PCIE_IO_PHYS_OFF +
+                   CONFIG_SYS_LS_PCIE_IO_SIZE))) {
     return EFI_INVALID_PARAMETER;
   }
